+++ grafanalib/elasticsearch.py
@@ -19,17 +19,25 @@ class CountMetricAgg(object):
     It's the default aggregator for elasticsearch queries.
     :param hide: show/hide the metric in the final panel display
     :param id: id of the metric
+    :param inline: script to apply to the data, using '_value'
     """
     id = attr.ib(default=0, validator=instance_of(int))
     hide = attr.ib(default=False, validator=instance_of(bool))
+    inline = attr.ib(default="", validator=instance_of(str))
 
     def to_json_data(self):
+        self.settings = {}
+
+        if self.inline:
+            self.settings['script'] = {'inline': self.inline}
+
         return {
             'id': str(self.id),
             'hide': self.hide,
             'type': 'count',
             'field': 'select field',
-            'settings': {},
+            'inlineScript': self.inline,
+            'settings': self.settings,
         }
 
 
@@ -42,18 +50,26 @@ class MaxMetricAgg(object):
     :param field: name of elasticsearch field to provide the maximum for
     :param hide: show/hide the metric in the final panel display
     :param id: id of the metric
+    :param inline: script to apply to the data, using '_value'
     """
     field = attr.ib(default="", validator=instance_of(str))
     id = attr.ib(default=0, validator=instance_of(int))
     hide = attr.ib(default=False, validator=instance_of(bool))
+    inline = attr.ib(default="", validator=instance_of(str))
 
     def to_json_data(self):
+        self.settings = {}
+
+        if self.inline:
+            self.settings['script'] = {'inline': self.inline}
+
         return {
             'id': str(self.id),
             'hide': self.hide,
             'type': 'max',
             'field': self.field,
-            'settings': {},
+            'inlineScript': self.inline,
+            'settings': self.settings,
         }
 
 
@@ -66,18 +82,26 @@ class CardinalityMetricAgg(object):
     :param field: name of elasticsearch field to provide the maximum for
     :param id: id of the metric
     :param hide: show/hide the metric in the final panel display
+    :param inline: script to apply to the data, using '_value'
     """
     field = attr.ib(default="", validator=instance_of(str))
     id = attr.ib(default=0, validator=instance_of(int))
     hide = attr.ib(default=False, validator=instance_of(bool))
+    inline = attr.ib(default="", validator=instance_of(str))
 
     def to_json_data(self):
+        self.settings = {}
+
+        if self.inline:
+            self.settings['script'] = {'inline': self.inline}
+
         return {
             'id': str(self.id),
             'hide': self.hide,
             'type': 'cardinality',
             'field': self.field,
-            'settings': {},
+            'inlineScript': self.inline,
+            'settings': self.settings,
         }
 
 
@@ -90,19 +114,27 @@ class AverageMetricAgg(object):
     :param field: name of elasticsearch field to provide the maximum for
     :param id: id of the metric
     :param hide: show/hide the metric in the final panel display
+    :param inline: script to apply to the data, using '_value'
     """
 
     field = attr.ib(default="", validator=instance_of(str))
     id = attr.ib(default=0, validator=instance_of(int))
     hide = attr.ib(default=False, validator=instance_of(bool))
+    inline = attr.ib(default="", validator=instance_of(str))
 
     def to_json_data(self):
+        self.settings = {}
+
+        if self.inline:
+            self.settings['script'] = {'inline': self.inline}
+
         return {
             'id': str(self.id),
             'hide': self.hide,
             'type': 'avg',
             'field': self.field,
-            'settings': {},
+            'inlineScript': self.inline,
+            'settings': self.settings,
             'meta': {}
         }
 
@@ -147,18 +179,26 @@ class SumMetricAgg(object):
     :param field: name of elasticsearch field to provide the sum over
     :param hide: show/hide the metric in the final panel display
     :param id: id of the metric
+    :param inline: script to apply to the data, using '_value'
     """
     field = attr.ib(default="", validator=instance_of(str))
     id = attr.ib(default=0, validator=instance_of(int))
     hide = attr.ib(default=False, validator=instance_of(bool))
+    inline = attr.ib(default="", validator=instance_of(str))
 
     def to_json_data(self):
+        self.settings = {}
+
+        if self.inline:
+            self.settings['script'] = {'inline': self.inline}
+
         return {
-            'type': 'sum',
             'id': str(self.id),
             'hide': self.hide,
+            'type': 'sum',
             'field': self.field,
-            'settings': {},
+            'inlineScript': self.inline,
+            'settings': self.settings,
         }
 
 
@@ -221,6 +261,7 @@ class BucketScriptAgg(object):
             })
 
         return {
+            'field': 'select field',
             'type': 'bucket_script',
             'id': str(self.id),
             'hide': self.hide,